prompt

+++ add_profile.ps1
@@ -1,5 +1,8 @@
-$d = Split-Path -parent $profile
-cp ./Microsoft.PowerShell_profile.ps1 -Destination $d
+$confirmation = Read-Host "Do you want to upload a new profile? Old profile will be overwritten! (y/n)"
+if ($confirmation -eq 'y') {
+    $d = Split-Path -parent $profile
+	cp ./Microsoft.PowerShell_profile.ps1 -Destination $d
 
-echo "New profile copied:"
-cat $profile
+	echo "New profile copied:"
+	cat $profile
+}
